+++ config/config.go
@@ -47,6 +47,10 @@ type Config struct {
 		}
 	}
 
+	SystemServices struct {
+		RelativePath string
+	}
+
 	Docker struct {
 		Socket string
 		Core   struct {
@@ -70,6 +74,7 @@ func New() (*Config, error) {
 	c.Core.Image = "mesg/core:" + strings.Split(version.Version, " ")[0]
 	c.Core.Name = "core"
 	c.Core.Path = filepath.Join(home, ".mesg")
+	c.SystemServices.RelativePath = "systemservices"
 	c.Core.Database.ServiceRelativePath = filepath.Join("database", "services")
 	c.Core.Database.ExecutionRelativePath = filepath.Join("database", "executions")
 	c.Docker.Core.Path = "/mesg"
@@ -114,8 +119,8 @@ func (c *Config) Prepare() error {
 
 // Validate checks values and return an error if any validation failed.
 func (c *Config) Validate() error {
-	if xstrings.SliceContains([]string{"text", "json"}, c.Log.Format) == false {
-		return fmt.Errorf("Value %q is not an allowed", c.Log.Format)
+	if !xstrings.SliceContains([]string{"text", "json"}, c.Log.Format) {
+		return fmt.Errorf("value %q is not an allowed", c.Log.Format)
 	}
 	if _, err := logrus.ParseLevel(c.Log.Level); err != nil {
 		return err