+++ database/service_db.go
@@ -4,19 +4,43 @@ import (
 	"encoding/json"
 	"errors"
 	"fmt"
+	"strings"
 
 	"github.com/mesg-foundation/core/service"
 	"github.com/sirupsen/logrus"
 	"github.com/syndtr/goleveldb/leveldb"
+	"github.com/syndtr/goleveldb/leveldb/util"
+)
+
+const (
+	sidKeyPrefix  = "sid_"
+	hashKeyPrefix = "hash_"
+)
+
+var (
+	errCannotSaveWithoutHash = errors.New("database: can't save service without hash")
+	errCannotSaveWithoutSID  = errors.New("database: can't save service without sid")
+	errSIDSameLen            = errors.New("database: sid can't have the same length as hash")
 )
 
 // ServiceDB describes the API of database package.
 type ServiceDB interface {
-	All() ([]*service.Service, error)
-	Close() error
-	Delete(id string) error
-	Get(id string) (*service.Service, error)
+	// Save saves a service to database.
 	Save(s *service.Service) error
+
+	// Get gets a service from database by its unique id
+	// or unique sid.
+	Get(hashOrSID string) (*service.Service, error)
+
+	// Delete deletes a service from database by its unique id
+	// or unique sid.
+	Delete(hashOrSID string) error
+
+	// All returns all services from database.
+	All() ([]*service.Service, error)
+
+	// Close closes underlying database connection.
+	Close() error
 }
 
 // LevelDBServiceDB is a database for storing service definition.
@@ -30,7 +54,6 @@ func NewServiceDB(path string) (*LevelDBServiceDB, error) {
 	if err != nil {
 		return nil, err
 	}
-
 	return &LevelDBServiceDB{db: db}, nil
 }
 
@@ -52,11 +75,11 @@ func (d *LevelDBServiceDB) unmarshal(id string, value []byte) (*service.Service,
 func (d *LevelDBServiceDB) All() ([]*service.Service, error) {
 	var (
 		services []*service.Service
-		iter     = d.db.NewIterator(nil, nil)
+		iter     = d.db.NewIterator(util.BytesPrefix([]byte(hashKeyPrefix)), nil)
 	)
-
 	for iter.Next() {
-		s, err := d.unmarshal(string(iter.Key()), iter.Value())
+		id := strings.TrimPrefix(string(iter.Key()), hashKeyPrefix)
+		s, err := d.unmarshal(id, iter.Value())
 		if err != nil {
 			// NOTE: Ignore all decode errors (possibly due to a service
 			// structure change or database corruption)
@@ -64,47 +87,123 @@ func (d *LevelDBServiceDB) All() ([]*service.Service, error) {
 				logrus.WithField("service", decodeErr.ID).Warning(decodeErr.Error())
 				continue
 			}
+			iter.Release()
 			return nil, err
 		}
 		services = append(services, s)
 	}
 	iter.Release()
-	if err := iter.Error(); err != nil {
-		return nil, err
-	}
-
-	return services, nil
+	return services, iter.Error()
 }
 
 // Delete deletes service from database.
-func (d *LevelDBServiceDB) Delete(id string) error {
-	return d.db.Delete([]byte(id), nil)
+func (d *LevelDBServiceDB) Delete(hashOrSID string) error {
+	tx, err := d.db.OpenTransaction()
+	if err != nil {
+		return err
+	}
+	if err := d.delete(tx, hashOrSID); err != nil {
+		tx.Discard()
+		return err
+	}
+	return tx.Commit()
+}
+
+// delete deletes service from database by using r reader.
+func (d *LevelDBServiceDB) delete(tx *leveldb.Transaction, hashOrSID string) error {
+	s, err := d.get(tx, hashOrSID)
+	if err != nil {
+		return err
+	}
+	if err := tx.Delete([]byte(hashKeyPrefix+s.Hash), nil); err != nil {
+		return err
+	}
+	return tx.Delete([]byte(sidKeyPrefix+s.SID), nil)
 }
 
 // Get retrives service from database.
-func (d *LevelDBServiceDB) Get(id string) (*service.Service, error) {
-	b, err := d.db.Get([]byte(id), nil)
+func (d *LevelDBServiceDB) Get(hashOrSID string) (*service.Service, error) {
+	tx, err := d.db.OpenTransaction()
+	if err != nil {
+		return nil, err
+	}
+	s, err := d.get(tx, hashOrSID)
+	if err != nil {
+		tx.Discard()
+		return nil, err
+	}
+	return s, tx.Commit()
+}
+
+// get retrives service from database by using r reader.
+func (d *LevelDBServiceDB) get(r leveldb.Reader, hashOrSID string) (*service.Service, error) {
+	hash := hashOrSID
+
+	// check if key is a sid, if yes then get hash.
+	bid, err := r.Get([]byte(sidKeyPrefix+hashOrSID), nil)
+	if err != nil && err != leveldb.ErrNotFound {
+		return nil, err
+	} else if err == nil {
+		hash = string(bid)
+	}
+
+	// get the service
+	b, err := r.Get([]byte(hashKeyPrefix+hash), nil)
 	if err != nil {
 		if err == leveldb.ErrNotFound {
-			return nil, &ErrNotFound{ID: id}
+			return nil, &ErrNotFound{ID: hashOrSID}
 		}
 		return nil, err
 	}
-
-	return d.unmarshal(id, b)
+	return d.unmarshal(hashOrSID, b)
 }
 
 // Save stores service in database.
+// If there is an another service that uses the same sid, it'll be deleted.
 func (d *LevelDBServiceDB) Save(s *service.Service) error {
-	if s.ID == "" {
-		return errors.New("database: can't save service without id")
+	// check service
+	if s.Hash == "" {
+		return errCannotSaveWithoutHash
 	}
-	b, err := d.marshal(s)
+	if s.SID == "" {
+		return errCannotSaveWithoutSID
+	}
+	if len(s.Hash) == len(s.SID) {
+		return errSIDSameLen
+	}
+
+	// open database transaction
+	tx, err := d.db.OpenTransaction()
 	if err != nil {
 		return err
 	}
 
-	return d.db.Put([]byte(s.ID), b, nil)
+	// delete existent service that has the same sid.
+	if err := d.delete(tx, s.SID); err != nil && !IsErrNotFound(err) {
+		tx.Discard()
+		return err
+	}
+
+	// encode service
+	b, err := d.marshal(s)
+	if err != nil {
+		tx.Discard()
+		return err
+	}
+
+	// save service with hash.
+	if err := tx.Put([]byte(hashKeyPrefix+s.Hash), b, nil); err != nil {
+		tx.Discard()
+		return err
+	}
+
+	// save sid-hash pair of service.
+	if err := tx.Put([]byte(sidKeyPrefix+s.SID), []byte(s.Hash), nil); err != nil {
+		tx.Discard()
+		return err
+	}
+
+	return tx.Commit()
 }
 
 // Close closes database.
@@ -127,10 +226,10 @@ type DecodeError struct {
 }
 
 func (e *DecodeError) Error() string {
-	return fmt.Sprintf("Database services: Could not decode service %q", e.ID)
+	return fmt.Sprintf("database: could not decode service %q", e.ID)
 }
 
-// IsErrNotFound returs true if err is type of ErrNotFound, false otherwise.
+// IsErrNotFound returns true if err is type of ErrNotFound, false otherwise.
 func IsErrNotFound(err error) bool {
 	_, ok := err.(*ErrNotFound)
 	return ok