+++ interface/grpc/core/stop_test.go
@@ -19,13 +19,14 @@ func TestStopService(t *testing.T) {
 	s, validationErr, err := server.api.DeployService(serviceTar(t, eventServicePath))
 	require.Zero(t, validationErr)
 	require.NoError(t, err)
-	defer server.api.DeleteService(s.ID)
+	defer server.api.DeleteService(s.Hash, false)
 
-	require.NoError(t, server.api.StartService(s.ID))
+	require.NoError(t, server.api.StartService(s.Hash))
 
 	reply, err := server.StopService(context.Background(), &coreapi.StopServiceRequest{
-		ServiceID: s.ID,
+		ServiceID: s.Hash,
 	})
+	require.NoError(t, err)
 
 	status, err := s.Status()
 	require.NoError(t, err)