+++ service/stop_test.go
@@ -23,7 +23,7 @@ func TestStopRunningService(t *testing.T) {
 
 	d, _ := s.getDependency(dependencyKey)
 
-	mc.On("Status", d.namespace()).Twice().Return(container.RUNNING, nil)
+	mc.On("Status", d.namespace()).Once().Return(container.RUNNING, nil)
 	mc.On("StopService", d.namespace()).Once().Return(nil)
 	mc.On("DeleteNetwork", s.namespace(), container.EventDestroy).Once().Return(nil)
 
@@ -33,11 +33,11 @@ func TestStopRunningService(t *testing.T) {
 	mc.AssertExpectations(t)
 }
 
-func TestStopNonRunningService(t *testing.T) {
+func TestStopDependency(t *testing.T) {
 	var (
 		dependencyKey = "1"
 		s, mc         = newFromServiceAndContainerMocks(t, &Service{
-			Name: "TestStopNonRunningService",
+			Name: "TestStopService",
 			Dependencies: []*Dependency{
 				{
 					Key:   dependencyKey,
@@ -48,60 +48,7 @@ func TestStopNonRunningService(t *testing.T) {
 	)
 
 	d, _ := s.getDependency(dependencyKey)
-
-	mc.On("Status", d.namespace()).Once().Return(container.STOPPED, nil)
-
-	err := s.Stop()
-	require.NoError(t, err)
-
-	mc.AssertExpectations(t)
-}
-
-func TestStopRunningDependency(t *testing.T) {
-	var (
-		dependencyKey = "1"
-		s, mc         = newFromServiceAndContainerMocks(t, &Service{
-			Name: "TestStopNonRunningService",
-			Dependencies: []*Dependency{
-				{
-					Key:   dependencyKey,
-					Image: "http-server",
-				},
-			},
-		})
-	)
-
-	d, _ := s.getDependency(dependencyKey)
-
-	mc.On("Status", d.namespace()).Once().Return(container.RUNNING, nil)
 	mc.On("StopService", d.namespace()).Once().Return(nil)
-
-	err := d.Stop()
-	require.NoError(t, err)
-
-	mc.AssertExpectations(t)
-}
-
-func TestStopNonRunningDependency(t *testing.T) {
-	var (
-		dependencyKey = "1"
-		s, mc         = newFromServiceAndContainerMocks(t, &Service{
-			Name: "TestStopNonRunningService",
-			Dependencies: []*Dependency{
-				{
-					Key:   dependencyKey,
-					Image: "http-server",
-				},
-			},
-		})
-	)
-
-	d, _ := s.getDependency(dependencyKey)
-
-	mc.On("Status", d.namespace()).Once().Return(container.STOPPED, nil)
-
-	err := d.Stop()
-	require.NoError(t, err)
-
+	require.NoError(t, d.Stop())
 	mc.AssertExpectations(t)
 }