+++ utils/workflowparser/definition.go
@@ -1,0 +1,5 @@
+package workflowparser
+
+// WorkflowDefinition TODO.
+type WorkflowDefinition struct {
+}